Update Numark-Mixtrack-3-scripts.js
Change PFL LED behaviour so that shift+Load (fader start) does not enable it to avoid confusing behaviour. Works properly with the smartPFL variable. The only thing I can't figure it out is that with these changes, the PFL button acts as a radio button - you can only choose to have PFL enabled on one deck but not both. Not necessarily an issue but not consistent with other mappings.